refactor(core): move BatchAuthTraceWriter beside prune_trace_files
The `.trace` credential-isolation contract was split across two crates: core
owned the retention half (`prune_trace_files`, moved in P1a) while src-tauri
still owned the writer that creates the files it prunes. Both halves encode the
same naming scheme — `batch-auth-<ts>.trace`, deliberately not `.log` and not
`tyutool-`-prefixed so no export or archive zip can ever collect it — so one
could have been changed without the other.

The writer is 40 lines over `std::fs::File` with no Tauri dependency, so this
is a move, not a rewrite. Its test moves with it.

Side effect worth noting for the next stage: `batch.rs` now has no reference to
`src-tauri`'s `logs` module at all, which removes one of the two things tying
`run_batch_auth_slot` to this crate.
